+++ src/machinable/storage/component.py
@@ -2,7 +2,7 @@ import os
 from typing import Union
 
 import pendulum
-
+import jsonlines
 from ..config.mapping import config_map
 from ..filesystem import open_fs, parse_storage_url
 from ..storage.models import StorageComponentModel
@@ -19,11 +19,12 @@ class StorageComponent:
         self._cache = cache or {}
         self._cache["experiment"] = experiment
 
-    def read_file(self, filepath, default=sentinel, reload=None):
+    def file(self, filepath, default=sentinel, reload=None):
         """Returns the content of a file in the storage
 
         # Arguments
         filepath: Relative filepath
+        default: Optional default if file does not exist
         reload: If True, cache will be ignored. If datetime, file will be reloaded
                 if cached version is older than the date
         """
@@ -91,7 +92,7 @@ class StorageComponent:
 
         return self._cache["experiment"]
 
-    def read_data(self, name=None, default=sentinel):
+    def data(self, name=None, default=sentinel):
         """Retrieves a data object from the storage
 
         # Arguments
@@ -101,22 +102,20 @@ class StorageComponent:
             with open_fs(self.url) as filesystem:
                 return filesystem.listdir("data")
 
-        return self.read_file(os.path.join("data", name), default)
+        return self.file(os.path.join("data", name), default)
 
     @property
     def config(self):
         """Returns the component config"""
         if "config" not in self._cache:
-            self._cache["config"] = config_map(
-                self.read_file("component.json")["config"]
-            )
+            self._cache["config"] = config_map(self.file("component.json")["config"])
         return self._cache["config"]
 
     @property
     def flags(self):
         """Returns the component flags"""
         if "flags" not in self._cache:
-            self._cache["flags"] = config_map(self.read_file("component.json")["flags"])
+            self._cache["flags"] = config_map(self.file("component.json")["flags"])
         return self._cache["flags"]
 
     @property
@@ -128,7 +127,7 @@ class StorageComponent:
     def components(self):
         if "components" not in self._cache:
             self._cache["components"] = [
-                config_map(component) for component in self.read_file("components.json")
+                config_map(component) for component in self.file("components.json")
             ]
 
         return self._cache["components"]
@@ -137,7 +136,7 @@ class StorageComponent:
     def host(self):
         """Returns information of the host"""
         if "host" not in self._cache:
-            self._cache["host"] = config_map(self.read_file("host.json"))
+            self._cache["host"] = config_map(self.file("host.json"))
         return self._cache["host"]
 
     @property
@@ -146,7 +145,7 @@ class StorageComponent:
         if "state" in self._cache:
             return self._cache["state"]
 
-        state = config_map(self.read_file("state.json"))
+        state = config_map(self.file("state.json"))
         if self.is_finished():
             self._cache["state"] = state
 
@@ -157,7 +156,7 @@ class StorageComponent:
         if "log" in self._cache:
             return self._cache["log"]
 
-        log = self.read_file("log.txt")
+        log = self.file("log.txt")
 
         if self.is_finished():
             self._cache["log"] = log
@@ -169,7 +168,7 @@ class StorageComponent:
         if "output" in self._cache:
             return self._cache["output"]
 
-        output = self.read_file("output.log")
+        output = self.file("output.log")
 
         if self.is_finished():
             self._cache["output"] = output
@@ -184,7 +183,7 @@ class StorageComponent:
     def has_records(self, scope="default"):
         """Returns True if records of given scope exist"""
         with open_fs(self.url) as filesystem:
-            return filesystem.exists(f"records/{scope}.json")
+            return filesystem.exists(f"records/{scope}.jsonl")
 
     def get_records(self, scope=None):
         """Returns a record collection
@@ -198,14 +197,26 @@ class StorageComponent:
             try:
                 with open_fs(self.url) as filesystem:
                     scopes = filesystem.listdir("records")
-                    return [s[:-5] for s in scopes if s.endswith(".json")]
+                    return [s[:-6] for s in scopes if s.endswith(".jsonl")]
             except FileNotFoundError:
                 return []
 
         if "records." + scope in self._cache:
             return self._cache["records." + scope]
 
-        records = RecordCollection(self.read_file(f"records/{scope}.p"))
+        records = []
+        try:
+            with open_fs(self.url) as filesystem:
+                if filesystem.isfile(f"records/{scope}.jsonl"):
+                    with jsonlines.Reader(
+                        filesystem.open(f"records/{scope}.jsonl")
+                    ) as reader:
+                        for record in reader.iter():
+                            # schema
+                            records.append(record)
+        except FileNotFoundError:
+            pass
+        records = RecordCollection(records)
 
         if self.is_finished():
             self._cache["records." + scope] = records